requirements.txt
python-dotenv
crewai
crewai_tools
langchain_community
langchain
crewai[tools]
langsmith
streamlit
pandas
openai